Move backfill logic to frontend
