+++ service/frontend/admin_handler_test.go
@@ -2061,30 +2061,30 @@ func (s *adminHandlerSuite) TestImportWorkflowExecution_WithNonAliasedSearchAttr
 				return "", serviceerror.NewInvalidArgument("unknown alias")
 			}).Times(eventsWithSasCount)
 
-		// Mock GetAlias - for valid field names it returns an alias, for invalid ones it returns an error.
-		if subTest.ExpectedErr != nil {
-			s.mockSaMapper.EXPECT().GetAlias(gomock.Any(), tv.NamespaceName().String()).Return("", serviceerror.NewInvalidArgument("")).AnyTimes()
-		} else {
-			// For valid field names, GetAlias returns an alias (field exists in the ES schema).
-			s.mockSaMapper.EXPECT().GetAlias(gomock.Any(), tv.NamespaceName().String()).Return("CustomKeywordField", nil).AnyTimes()
-			s.mockVisibilityMgr.EXPECT().ValidateCustomSearchAttributes(gomock.Any()).Return(nil, nil).Times(eventsWithSasCount)
-			s.mockHistoryClient.EXPECT().ImportWorkflowExecution(gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, request *historyservice.ImportWorkflowExecutionRequest, opts ...grpc.CallOption) (*historyservice.ImportWorkflowExecutionResponse, error) {
-				s.Equal(tv.NamespaceID().String(), request.NamespaceId)
-				for _, historyBatch := range request.HistoryBatches {
-					events, err := serializer.DeserializeEvents(historyBatch)
-					s.NoError(err)
-					for _, event := range events {
-						unaliasedSas, eventHasSas := searchattribute.GetFromEvent(event)
-						if eventHasSas {
-							s.NotNil(unaliasedSas, "search attributes must be set on every event with search_attributes field")
-							s.Len(unaliasedSas.GetIndexedFields(), 1, "only 1 search attribute must be set")
-							s.ProtoEqual(saValue, unaliasedSas.GetIndexedFields()["Keyword01"])
+			// Mock GetAlias - for valid field names it returns an alias, for invalid ones it returns an error.
+			if subTest.ExpectedErr != nil {
+				s.mockSaMapper.EXPECT().GetAlias(gomock.Any(), tv.NamespaceName().String()).Return("", serviceerror.NewInvalidArgument("")).AnyTimes()
+			} else {
+				// For valid field names, GetAlias returns an alias (field exists in the ES schema).
+				s.mockSaMapper.EXPECT().GetAlias(gomock.Any(), tv.NamespaceName().String()).Return("CustomKeywordField", nil).AnyTimes()
+				s.mockVisibilityMgr.EXPECT().ValidateCustomSearchAttributes(gomock.Any()).Return(nil, nil).Times(eventsWithSasCount)
+				s.mockHistoryClient.EXPECT().ImportWorkflowExecution(gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, request *historyservice.ImportWorkflowExecutionRequest, opts ...grpc.CallOption) (*historyservice.ImportWorkflowExecutionResponse, error) {
+					s.Equal(tv.NamespaceID().String(), request.NamespaceId)
+					for _, historyBatch := range request.HistoryBatches {
+						events, err := serializer.DeserializeEvents(historyBatch)
+						s.NoError(err)
+						for _, event := range events {
+							unaliasedSas, eventHasSas := searchattribute.GetFromEvent(event)
+							if eventHasSas {
+								s.NotNil(unaliasedSas, "search attributes must be set on every event with search_attributes field")
+								s.Len(unaliasedSas.GetIndexedFields(), 1, "only 1 search attribute must be set")
+								s.ProtoEqual(saValue, unaliasedSas.GetIndexedFields()["Keyword01"])
+							}
 						}
 					}
-				}
-				return &historyservice.ImportWorkflowExecutionResponse{}, nil
-			})
-		}
+					return &historyservice.ImportWorkflowExecutionResponse{}, nil
+				})
+			}
 
 			_, err := s.handler.ImportWorkflowExecution(context.Background(), &adminservice.ImportWorkflowExecutionRequest{
 				Namespace:      tv.NamespaceName().String(),
