+++ service/frontend/fx.go
@@ -1,6 +1,7 @@
 package frontend
 
 import (
+	"context"
 	"fmt"
 	"net"
 
@@ -115,6 +116,7 @@ var Module = fx.Options(
 	fx.Provide(NexusEndpointRegistryProvider),
 	fx.Invoke(ServiceLifetimeHooks),
 	fx.Invoke(EndpointRegistryLifetimeHooks),
+	fx.Invoke(SearchAttributeMappingsBackfillLifetimeHooks),
 	fx.Provide(schedulerpb.NewSchedulerServiceLayeredClient),
 	nexusfrontend.Module,
 	activity.FrontendModule,
@@ -1004,4 +1006,36 @@ func EndpointRegistryLifetimeHooks(lc fx.Lifecycle, registry nexus.EndpointRegis
 
 func ServiceLifetimeHooks(lc fx.Lifecycle, svc *Service) {
 	lc.Append(fx.StartStopHook(svc.Start, svc.Stop))
+}
+
+// SearchAttributeMappingsBackfillLifetimeHooks runs the search attribute backfill on frontend startup.
+// Only runs when Elasticsearch visibility is configured; updates go through the namespace handler.
+func SearchAttributeMappingsBackfillLifetimeHooks(
+	lc fx.Lifecycle,
+	cfg *config.Config,
+	clusterMetadata cluster.Metadata,
+	metadataManager persistence.MetadataManager,
+	clusterMetadataManager persistence.ClusterMetadataManager,
+	handler Handler,
+	logger log.Logger,
+) {
+	visDataStore := cfg.Persistence.GetVisibilityStoreConfig()
+	if visDataStore.Elasticsearch == nil {
+		return
+	}
+	visibilityIndexName := visDataStore.GetIndexName()
+	currentClusterName := clusterMetadata.GetCurrentClusterName()
+	lc.Append(fx.Hook{
+		OnStart: func(ctx context.Context) error {
+			return InitializeSearchAttributeMappings(
+				ctx,
+				metadataManager,
+				clusterMetadataManager,
+				currentClusterName,
+				visibilityIndexName,
+				handler,
+				logger,
+			)
+		},
+	})
 }
